teacher_llm/requirements.txt
fastapi
uvicorn
requests
python-dotenv
redis
ollama
pydantic
